solved critical and major issues + added a command to create new admin user

+++ app/Http/Controllers/Api/servicesController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Models\Service;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 
 class servicesController extends Controller
@@ -25,7 +26,10 @@ class servicesController extends Controller
     }
 
     public function getServices(Request $request){
-        $services=Service::with('plans')->get();
+        $services=Service::with(['plans'=>function($query){
+            $query->where('is_active',1);
+        }])->get();
+
         return response()->json([
             'message'=>'all services',
             'services'=>$services
@@ -60,6 +64,22 @@ class servicesController extends Controller
                 'message'=>'Service not Found'
             ],404);
         }
+        // check for active subscriptions
+        $activeSubscriptions=$service->subscriptions()->where('status','active')->where('ends_at','>',Carbon::now())->get();
+        if($activeSubscriptions->count()>0){
+            return response()->json([
+                'message'=>'Service has active subscriptions'
+            ],400);
+        }
+        // check for pending requests
+        $pendingRequests=$service->subscriptionRequests()->where('status','pending')->get();
+        if($pendingRequests->count()>0){
+            return response()->json([
+                'message'=>'Service has pending requests'
+            ],400);
+        }
+
+        // delete service if no active subscriptions and no pending requests
         $service->delete();
 
         return response()->json([
